+++ Libftprintf/libftprintf.h
@@ -6,7 +6,7 @@
 /*   By: lwiller <lwiller@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/04 06:59:10 by lwiller           #+#    #+#             */
-/*   Updated: 2021/01/08 11:35:28 by lwiller          ###   ########lyon.fr   */
+/*   Updated: 2021/01/12 16:47:29 by lwiller          ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,15 +20,14 @@ int		ft_atoi(const char *str);
 void	ft_bzero(void *s, size_t n);
 void	*ft_calloc(size_t count, size_t size);
 int		ft_isdigit(int c);
-char	*ft_itoa(long int n);
-char	*ft_iota_base(long int nb, char *base);
-int 	ft_putchar_int(char c);
+char	*ft_itoa(long long n);
+char	*ft_iota_base(unsigned long long nb, char *base);
+int		ft_putchar_int(char c);
 int		ft_putstr_int(char *str);
 char	*ft_strdup(char *str);
 char	*ft_strjoin(char *s1, char *s2);
 char	*ft_strjoin_left(char s1, char *s2);
 char	*ft_strjoin_right(char *s1, char s2);
-int 	ft_strlen(char *str);
-
+int		ft_strlen(char *str);
 
 #endif
